+++ CHANGELOG.md
@@ -11,6 +11,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Changed
+
+- Set SQLite temp store to memory.
+
 ### Fixed
 
 - Date param dependency causing re-run of query dailies.
@@ -18,6 +22,7 @@ All notable changes to this project will be documented in this file.
 ### Improved
 
 - Persisted computed values on points and removed view. Quest list query execution time reduced from 4~5s to <0.01s
+- Increased SQLite cache size.
 
 ## [v0.2.7]